Fixed issue with extra ampersand being included right after question mark.
An extra ampersand was being included after the question mark when appending query string parameters.

https://www.example.com?&parameter1=value1

This caused problems with requests to controller methods on the server that included optional parameters. The request did not get routed to the correct controller method and resulted in a 404.

Note that supplying values for the optional parameters was a workaround.